Moving top level files around.
Eventually, most of the top level files will be in a directory.
