+++ package.json
@@ -41,16 +41,17 @@
     "extends": "@istanbuljs/nyc-config-typescript"
   },
   "scripts": {
-    "build": "npm run lint && npm run compile && npm run templates && npm run webpack",
-    "clean": "rm -r ./build/ ./styles.css || true",
+    "build": "npm run lint && npm run makeJson && npm run compile && npm run templates && npm run webpack",
+    "clean": "rm -r ./build/ assets/settings.json assets/translations.json || true",
     "compile": "tsc",
     "cover": "nyc mocha --file build/tests/utils/Vue.js --recursive build/tests",
-    "fix": "eslint --fix src server.ts script.ts tests --ext ts",
-    "lint": "eslint src server.ts script.ts tests --ext ts",
+    "fix": "eslint --fix src tests --ext ts",
+    "lint": "eslint src tests --ext ts",
+    "makeJson": "node make_static_json.js",
     "precover": "tsc --build tsconfig-test.json",
-    "prebuild": "node make_static_json.js && lessc src/styles/common.less styles.css",
+    "prebuild": "mkdir -p build/ && lessc src/styles/common.less > build/styles.css",
     "pretest": "tsc --build tsconfig-test.json",
-    "start": "node build/server.js",
+    "start": "node build/src/server.js",
     "templates": "node compile-vue-templates.js && tsc --build tsconfig-vue.json",
     "test": "mocha --file build/tests/utils/Vue.js --recursive build/tests",
     "testgiven": "tsc --build tsconfig.json && npm run webpack && tsc --build tsconfig-test.json && mocha",
