+++ tsconfig.json
@@ -21,7 +21,5 @@
     },
     "include": [
         "src/**/*.ts",
-        "script.ts",
-        "server.ts"
     ]
 }
